Refactor directory structure: Move all resolvers to its separate  file
