+++ src/resolvers/User.js
@@ -1,0 +1,16 @@
+'use strict'
+
+const User = {
+  posts(parent, args, { db }, info) {
+    return db.posts.filter((post) => {
+      return post.author === parent.id
+    })
+  },
+  comments(parent, args, { db }, info) {
+    return db.comments.filter((comment) => {
+      return comment.author === parent.id
+    })
+  }
+}
+
+export default User
